Split sync payload collectors
